# backend/docs/swagger.yaml
definitions:
...
  handlers.JobApplicationWithApplicantDetails:
    properties:
      aboutMe:
        type: string
      birthDate:
        type: string
      createdAt:
        type: string
      email:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      github:
        type: string
      id:
        type: integer
      jobId:
        type: integer
      linkedIn:
        type: string
      major:
        type: string
      phone:
        type: string
      photoId:
        type: string
      status:
        $ref: '#/definitions/model.JobApplicationStatus'
      studentId:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
  handlers.JobApplicationWithApplicantName:
    properties:
      createdAt:
        type: string
      email:
        type: string
      files:
        items:
          $ref: '#/definitions/model.File'
        type: array
      id:
        type: integer
      jobId:
        type: integer
      major:
        type: string
      phone:
        type: string
      status:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
...
  handlers.UpdateJobApplicationStatusHandler.UpdateStatusInput:
    properties:
      status:
        enum:
        - accepted
        - rejected
        - pending
        type: string
    required:
    - status
    type: object
...
  model.File:
    properties:
      category:
        $ref: '#/definitions/model.FileCategory'
      createdAt:
        type: string
      fileType:
        $ref: '#/definitions/model.FileType'
      id:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  model.FileCategory:
    enum:
    - image
    - document
    type: string
    x-enum-varnames:
    - FileCategoryImage
    - FileCategoryDocument
  model.FileType:
    enum:
    - jpg
    - jpeg
    - png
    - webp
    - pdf
    - doc
    - docx
    type: string
    x-enum-varnames:
    - FileTypeJPG
    - FileTypeJPEG
    - FileTypePNG
    - FileTypeWEBP
    - FileTypePDF
    - FileTypeDOC
    - FileTypeDOCX
  model.JobApplicationStatus:
    enum:
    - accepted
    - rejected
    - pending
    type: string
    x-enum-varnames:
    - JobApplicationAccepted
    - JobApplicationRejected
    - JobApplicationPending
host: localhost:8000
info:
  contact: {}
  description: This is a sample API for KU-Work
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: KU-Work API
  version: "1.0"
paths:
  /applications:
    get:
      description: Fetches job applications for the authenticated user. If the user
        is a company, it returns all applications for all their job postings. If the
        user is a student, it returns all of their own applications. Supports pagination.
      parameters:
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 32
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of job applications
          schema:
            items:
              $ref: '#/definitions/handlers.JobApplicationWithApplicantName'
            type: array
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: 'Forbidden: User is not a company or student'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all applications for the current user
      tags:
      - Job Applications
...
  /jobs/{id}/applications:
    get:
      description: Fetches all job applications for a specific job posting. This endpoint
        is for companies to view applicants. It supports status filtering (pending,
        accepted, rejected) and pagination.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by status (pending, accepted, rejected)
        in: query
        name: status
        type: string
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 32
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of job applications
          schema:
            items:
              $ref: '#/definitions/handlers.JobApplicationWithApplicantName'
            type: array
        "400":
          description: 'Bad Request: Invalid job ID'
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: 'Forbidden: User is not authorized to view these applications'
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 'Not Found: Job not found'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get applications for a specific job
      tags:
      - Job Applications
  /jobs/{id}/applications/{studentId}:
    get:
      description: Retrieves detailed information about a single job application for
        a specific student, including the applicant's full profile, contact information,
        and attached files (resume, etc.).
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student User ID
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detailed job application
          schema:
            $ref: '#/definitions/handlers.JobApplicationWithApplicantDetails'
        "400":
          description: 'Bad Request: Invalid job ID'
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 'Not Found: Job application not found'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a specific job application
      tags:
      - Job Applications
  /jobs/{id}/applications/{studentId}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of a job application to 'accepted', 'rejected',
        or 'pending'. This action can only be performed by the company that posted
        the job.
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student User ID
        in: path
        name: studentId
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateJobApplicationStatusHandler.UpdateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: Application status updated successfully
          schema:
            properties:
              ' status':
                type: string
              message:
                type: string
            type: object
        "400":
          description: 'Bad Request: Invalid ID or input'
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: 'Forbidden: User is not authorized to update this application'
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 'Not Found: Job or application not found'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update job application status
      tags:
      - Job Applications
  /jobs/{id}/apply:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new job application. Allows an approved student to apply
        to an approved job posting by submitting their application with optional alternate
        contact information and required document files (e.g., resume, cover letter).
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - description: Files to upload (e.g., Resume, Cover Letter). Max 2 files.
        in: formData
        name: Files
        required: true
        type: file
      - description: Alternate phone number
        in: formData
        name: AltPhone
        type: string
      - description: Alternate email address
        in: formData
        name: AltEmail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created job application
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: 'Forbidden: Student status not approved'
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: 'Not Found: Invalid Job ID'
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Apply to a job
      tags:
      - Job Applications
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"